backend/pyproject.toml
[project]
name = "cyber-econ-backend"
version = "0.1.0"
description = "Cyber-Econ Backend API"
authors = [
    {name = "Antigravity", email = "antigravity@example.com"},
]
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "motor>=3.0.0",
    "pydantic>=2.0.0",
    "groq",
    "python-dotenv",
]
requires-python = ">=3.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "cyber-econ-backend"
version = "0.1.0"
description = "Cyber-Econ Backend API"
authors = ["Antigravity <antigravity@example.com>"]

[tool.poetry.dependencies]
python = ">=3.9"
fastapi = "^0.100.0"
uvicorn = "^0.20.0"
motor = "^3.0.0"
pydantic = "^2.0.0"
groq = "^0"
